@@ -4,27 +4,28 @@
   "private": true,
   "homepage": "https://ctmd.renci.org/",
   "dependencies": {
-    "@material-ui/core": "^4.7.2",
-    "@material-ui/icons": "^3.0.2",
-    "@material-ui/lab": "^4.0.0-alpha.35",
-    "@material-ui/styles": "^4.1.1",
-    "@nivo/bar": "^0.58.0",
-    "@nivo/calendar": "^0.58.0",
-    "@nivo/line": "^0.61.1",
-    "@nivo/pie": "^0.58.0",
-    "axios": "^0.21.2",
-    "caniuse-lite": "^1.0.30001015",
-    "d3": "^5.9.1",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@material-ui/styles": "^4.11.5",
+    "@nivo/bar": "^0.87.0",
+    "@nivo/calendar": "^0.87.0",
+    "@nivo/line": "^0.87.0",
+    "@nivo/pie": "^0.87.0",
+    "ajv": "^8.17.1",
+    "axios": "^1.7.9",
+    "caniuse-lite": "^1.0.30001686",
+    "d3": "^5.16.0",
     "d3-sankey": "^0.12.3",
-    "d3-scale-chromatic": "^1.3.3",
+    "d3-scale-chromatic": "^1.5.0",
     "d3-tip": "^0.9.1",
-    "material-table": "^1.54.1",
-    "react": "^16.8.6",
+    "material-table": "^1.69.3",
+    "react": "^18.3.1",
     "react-csv": "^2.2.2",
-    "react-dom": "^16.8.6",
-    "react-ga4": "^1.4.1",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "^2.1.8",
+    "react-dom": "^18.3.1",
+    "react-ga4": "^2.1.0",
+    "react-router-dom": "^6.28.0",
+    "react-scripts": "^5.0.1",
     "webfontloader": "^1.6.28"
   },
   "scripts": {
