@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App';
 import * as serviceWorker from './serviceWorker'
@@ -16,7 +16,10 @@ WebFont.load({
     }
 })
 
-ReactDOM.render(<App/>, document.getElementById('root'))
+// React 18 createRoot API
+const container = document.getElementById('root')
+const root = createRoot(container)
+root.render(<App />)
 
 // If you want your App to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
